Inheritance re-model to re-use base TCP functionality in the SSL class. Client and server still runs.

+++ IoServices/IoServices.cpp
@@ -112,7 +112,7 @@ namespace AsyncIo
 			auto lockedConn = weakStreamConn.lock();
 			if (lockedConn)
 			{
-				readCb(lockedConn, available);
+				readCb(lockedConn, static_cast<int>(available));
 			}
 			else
 			{
